+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Adicionado lib `react-countup` para animar um número de 0 a x.
+- Criado componente `LevelProgressBar` (https://github.com/3C-gg/reload-frontend/issues/137).
 - Criado `ArrowUpIcon` componente.
 - Criado componente `HeaderProfileMenu` usado no `Header` (https://github.com/3C-gg/reload-frontend/issues/130).
 - Criado componente `LinkButton` usado no `Header` que muda conforme o status do lobby e match (https://github.com/3C-gg/reload-frontend/issues/127).